modif

+++ TP1/prog1.py
@@ -4,8 +4,8 @@ import fonctions as f
 print("hello ,World!")
 
 
-chiffre1 = int(input("Entrez un chiffre : "))
-chiffre2 = int(input("Entrez un chiffre : "))
+chiffre1 = float(input("Entrez un chiffre : "))
+chiffre2 = float(input("Entrez un chiffre : "))
 
 res = f.puissance(chiffre1,chiffre2)
